+++ ImdP-Server/src/main/java/br/ufrn/imd/business/JustificativaFaltaService.java
@@ -1,9 +1,11 @@
 package br.ufrn.imd.business;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
+import javax.persistence.NoResultException;
 
 import br.ufrn.imd.dao.JustificativaFaltaDao;
 import br.ufrn.imd.dominio.JustificativaFalta;
@@ -11,13 +13,19 @@ import br.ufrn.imd.dominio.JustificativaFalta;
 @Stateless
 public class JustificativaFaltaService{
 
-	
 	@Inject
 	private JustificativaFaltaDao justificativaFaltaDao;
 	
 	public void save(JustificativaFalta justificativaFalta) {
 		verificarJustificativaFalta(justificativaFalta);
-		justificativaFaltaDao.save(justificativaFalta);
+		JustificativaFalta justficativa = justificativaFaltaDao.buscarPorId(justificativaFalta.getIdJustificativaFalta());
+		
+		if(justficativa == null){
+			justificativaFaltaDao.save(justificativaFalta);
+		}
+		else{
+			throw new IllegalArgumentException("Erro - save: JustificativaFalta ja existe na base de dados");
+		}
 	}
 
 	public JustificativaFalta update(JustificativaFalta justificativaFalta) {
@@ -26,15 +34,60 @@ public class JustificativaFaltaService{
 	}
 
 	public void delete(JustificativaFalta justificativaFalta) {
-		justificativaFaltaDao.delete(justificativaFalta);
+		verificarJustificativaFalta(justificativaFalta);
+		JustificativaFalta justficativa = justificativaFaltaDao.buscarPorId(justificativaFalta.getIdJustificativaFalta());
+		
+		if(justficativa == null){
+			throw new IllegalArgumentException("Erro - delete: Justificativa de falta nao existe na base de dados");
+		}
+		else{
+			justificativaFaltaDao.delete(justificativaFalta);
+		}
 	}
 
-	public JustificativaFalta find(int entityID) {
-		return (JustificativaFalta) justificativaFaltaDao.find(entityID);
+	public JustificativaFalta buscar(int entityID) {
+		JustificativaFalta justficativa = new JustificativaFalta();
+		justficativa = buscaId(entityID);
+		
+		if(justficativa == null){
+			throw new NoResultException("Erro - buscar: Justificativa de falta nao encontrada");
+		}
+		else{
+			return justficativa;
+		}
 	}
 
-	public List<JustificativaFalta> findAll() {
-		return justificativaFaltaDao.listar();
+	public List<JustificativaFalta> listar() {
+		ArrayList<JustificativaFalta> justficativas = new ArrayList<JustificativaFalta>();
+		justficativas = justificativaFaltaDao.listar();
+		if(justficativas.size() > 0){
+			return justficativas;
+		}
+		else{
+			throw new NoResultException("Erro - listar: nenhuma justificativa de falta cadastrado");
+		}
+	}
+	
+	public ArrayList<JustificativaFalta> buscarFiltro(String nomeUsuario, int idUnidade, int idSetor){
+		ArrayList<JustificativaFalta> justficativas = new ArrayList<JustificativaFalta>();
+		justficativas = justificativaFaltaDao.buscarJustificativaFiltro(nomeUsuario, idUnidade, idSetor);
+		if(justficativas.size() > 0){
+			return justficativas;
+		}
+		else{
+			return null;
+		}
+	}
+
+	//Metodo usados apenas para verificacoes no escopo do service
+	private JustificativaFalta buscaId(int id){
+		try{
+			JustificativaFalta justficativa = justificativaFaltaDao.buscarPorId(id);
+			return justficativa;
+		}
+		catch(NoResultException e){
+			return null;
+		}
 	}
 	
 	private void verificarJustificativaFalta(JustificativaFalta justificativaFalta){
@@ -61,8 +114,7 @@ public class JustificativaFaltaService{
 			hasError = true;
 		}
 		if (hasError){
-			throw new IllegalArgumentException("A justificativa de Falta nao possui todos os dados.");
+			throw new IllegalArgumentException("A justificativa de falta nao possui todos os dados.");
 		}
 	}
-	
 }
