+++ ImdP-Server/src/main/java/br/ufrn/imd/business/MesTrabalhoService.java
@@ -1,9 +1,11 @@
 package br.ufrn.imd.business;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
+import javax.persistence.NoResultException;
 
 import br.ufrn.imd.dao.MesTrabalhoDao;
 import br.ufrn.imd.dominio.MesTrabalho;
@@ -11,13 +13,19 @@ import br.ufrn.imd.dominio.MesTrabalho;
 @Stateless
 public class MesTrabalhoService{
 
-	
 	@Inject
 	private MesTrabalhoDao mesTrabalhoDao;
 	
 	public void save(MesTrabalho mesTrabalho) {
 		verificarMesTrabalho(mesTrabalho);
-		mesTrabalhoDao.save(mesTrabalho);
+		MesTrabalho month = mesTrabalhoDao.buscarPorId(mesTrabalho.getIdMesTrabalho());
+		
+		if(month == null){
+			mesTrabalhoDao.save(mesTrabalho);
+		}
+		else{
+			throw new IllegalArgumentException("Erro - save: MesTrabalho ja existe na base de dados");
+		}
 	}
 
 	public MesTrabalho update(MesTrabalho mesTrabalho) {
@@ -26,17 +34,62 @@ public class MesTrabalhoService{
 	}
 
 	public void delete(MesTrabalho mesTrabalho) {
-		mesTrabalhoDao.delete(mesTrabalho);
+		verificarMesTrabalho(mesTrabalho);
+		MesTrabalho month = mesTrabalhoDao.buscarPorId(mesTrabalho.getIdMesTrabalho());
+		
+		if(month == null){
+			throw new IllegalArgumentException("Erro - delete: MesTrabalho nao existe na base de dados");
+		}
+		else{
+			mesTrabalhoDao.delete(mesTrabalho);
+		}
 	}
 
-	public MesTrabalho find(int entityID) {
-		return (MesTrabalho) mesTrabalhoDao.find(entityID);
+	public MesTrabalho buscar(int entityID) {
+		MesTrabalho month = new MesTrabalho();
+		month = buscaId(entityID);
+		
+		if(month == null){
+			throw new NoResultException("Erro - buscar: MesTrabalho nao encontrado");
+		}
+		else{
+			return month;
+		}
 	}
 
-	public List<MesTrabalho> findAll() {
-		return mesTrabalhoDao.listar();
+	public List<MesTrabalho> listar() {
+		ArrayList<MesTrabalho> months = new ArrayList<MesTrabalho>();
+		months = mesTrabalhoDao.listar();
+		if(months.size() > 0){
+			return months;
+		}
+		else{
+			throw new NoResultException("Erro - listar: nenhum mesTrabalho cadastrado");
+		}
 	}
 	
+	public ArrayList<MesTrabalho> buscarFiltro(int mes, int ano){
+		ArrayList<MesTrabalho> months = new ArrayList<MesTrabalho>();
+		months = mesTrabalhoDao.buscarMesTrabalhoFiltro(mes, ano);
+		if(months.size() > 0){
+			return months;
+		}
+		else{
+			return null;
+		}
+	}
+
+	//Metodo usados apenas para verificacoes no escopo do service
+	private MesTrabalho buscaId(int id){
+		try{
+			MesTrabalho month = mesTrabalhoDao.buscarPorId(id);
+			return month;
+		}
+		catch(NoResultException e){
+			return null;
+		}
+	}
+		
 	private void verificarMesTrabalho(MesTrabalho mesTrabalho){
 		boolean hasError = false;
 		
@@ -60,5 +113,4 @@ public class MesTrabalhoService{
 			throw new IllegalArgumentException("O mesTrabalho nao possui todos os dados.");
 		}
 	}
-	
 }
