+++ ImdP-Server/src/main/java/br/ufrn/imd/business/TipoJustificativaService.java
@@ -1,9 +1,11 @@
 package br.ufrn.imd.business;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
+import javax.persistence.NoResultException;
 
 import br.ufrn.imd.dao.TipoJustificativaDao;
 import br.ufrn.imd.dominio.TipoJustificativa;
@@ -16,7 +18,14 @@ public class TipoJustificativaService{
 	
 	public void save(TipoJustificativa tipoJustificativa) {
 		verificarTipoJustificativa(tipoJustificativa);
-		tipoJustificativaDao.save(tipoJustificativa);
+		TipoJustificativa tj = tipoJustificativaDao.buscarPorId(tipoJustificativa.getIdTipoJustificativa());
+		
+		if(tj == null){
+			tipoJustificativaDao.save(tipoJustificativa);
+		}
+		else{
+			throw new IllegalArgumentException("Erro - save: tipo de justificativa ja existe na base de dados");
+		}
 	}
 
 	public TipoJustificativa update(TipoJustificativa tipoJustificativa) {
@@ -25,17 +34,62 @@ public class TipoJustificativaService{
 	}
 
 	public void delete(TipoJustificativa tipoJustificativa) {
-		tipoJustificativaDao.delete(tipoJustificativa);
+		verificarTipoJustificativa(tipoJustificativa);
+		TipoJustificativa tj = tipoJustificativaDao.buscarPorId(tipoJustificativa.getIdTipoJustificativa());
+		
+		if(tj == null){
+			throw new IllegalArgumentException("Erro - delete: tipo de justificativa nao existe na base de dados");
+		}
+		else{
+			tipoJustificativaDao.delete(tipoJustificativa);
+		}
 	}
 
-	public TipoJustificativa find(int entityID) {
-		return (TipoJustificativa) tipoJustificativaDao.find(entityID);
+	public TipoJustificativa buscar(int entityID) {
+		TipoJustificativa tj = new TipoJustificativa();
+		tj = buscaId(entityID);
+		
+		if(tj == null){
+			throw new NoResultException("Erro - buscar: tipo de justificativa nao encontrado");
+		}
+		else{
+			return tj;
+		}
 	}
 
-	public List<TipoJustificativa> findAll() {
-		return tipoJustificativaDao.listar();
+	public List<TipoJustificativa> listar() {
+		ArrayList<TipoJustificativa> tjs = new ArrayList<TipoJustificativa>();
+		tjs = tipoJustificativaDao.listar();
+		if(tjs.size() > 0){
+			return tjs;
+		}
+		else{
+			throw new NoResultException("Erro - listar: nenhum tipo de justificativa cadastrado");
+		}
 	}
 	
+	public ArrayList<TipoJustificativa> buscarFiltro(String nomeTipoJustificativa){
+		ArrayList<TipoJustificativa> tjs = new ArrayList<TipoJustificativa>();
+		tjs = tipoJustificativaDao.buscarTipoJustificativaFiltro(nomeTipoJustificativa);
+		if(tjs.size() > 0){
+			return tjs;
+		}
+		else{
+			return null;
+		}
+	}
+
+	//Metodo usados apenas para verificacoes no escopo do service
+	private TipoJustificativa buscaId(int id){
+		try{
+			TipoJustificativa tj = tipoJustificativaDao.buscarPorId(id);
+			return tj;
+		}
+		catch(NoResultException e){
+			return null;
+		}
+	}
+
 	private void verificarTipoJustificativa(TipoJustificativa tipoJustificativa){
 		boolean hasError = false;
 		
@@ -49,5 +103,4 @@ public class TipoJustificativaService{
 			throw new IllegalArgumentException("O tipo de justificativa nao possui todos os dados.");
 		}
 	}
-	
 }
