+++ ImdP-Server/src/main/java/br/ufrn/imd/business/UnidadeService.java
@@ -1,9 +1,11 @@
 package br.ufrn.imd.business;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import javax.ejb.Stateless;
 import javax.inject.Inject;
+import javax.persistence.NoResultException;
 
 import br.ufrn.imd.dao.UnidadeDao;
 import br.ufrn.imd.dominio.Unidade;
@@ -16,7 +18,14 @@ public class UnidadeService{
 	
 	public void save(Unidade unidade) {
 		verificarUnidade(unidade);
-		unidadeDao.save(unidade);
+		Unidade unit = unidadeDao.buscarPorId(unidade.getIdUnidade());
+		
+		if(unit == null){
+			unidadeDao.save(unidade);
+		}
+		else{
+			throw new IllegalArgumentException("Erro - save: Unidade ja existe na base de dados");
+		}
 	}
 
 	public Unidade update(Unidade unidade) {
@@ -25,15 +34,60 @@ public class UnidadeService{
 	}
 
 	public void delete(Unidade unidade) {
-		unidadeDao.delete(unidade);
+		verificarUnidade(unidade);
+		Unidade unit = unidadeDao.buscarPorId(unidade.getIdUnidade());
+		
+		if(unit == null){
+			throw new IllegalArgumentException("Erro - delete: Unidade nao existe na base de dados");
+		}
+		else{
+			unidadeDao.delete(unidade);
+		}
 	}
 
-	public Unidade find(int entityID) {
-		return (Unidade) unidadeDao.find(entityID);
+	public Unidade buscar(int entityID) {
+		Unidade unit = new Unidade();
+		unit = buscaId(entityID);
+		
+		if(unit == null){
+			throw new NoResultException("Erro - buscar: Unidade nao encontrada");
+		}
+		else{
+			return unit;
+		}
 	}
 
-	public List<Unidade> findAll() {
-		return unidadeDao.listar();
+	public List<Unidade> listar() {
+		ArrayList<Unidade> units = new ArrayList<Unidade>();
+		units = unidadeDao.listar();
+		if(units.size() > 0){
+			return units;
+		}
+		else{
+			throw new NoResultException("Erro - listar: nenhuma unidade cadastrada");
+		}
+	}
+	
+	public ArrayList<Unidade> buscarFiltro(String nomeUnidade){
+		ArrayList<Unidade> units = new ArrayList<Unidade>();
+		units = unidadeDao.buscarUnidadeFiltro(nomeUnidade);
+		if(units.size() > 0){
+			return units;
+		}
+		else{
+			return null;
+		}
+	}
+
+	//Metodo usados apenas para verificacoes no escopo do service
+	private Unidade buscaId(int id){
+		try{
+			Unidade unit = unidadeDao.buscarPorId(id);
+			return unit;
+		}
+		catch(NoResultException e){
+			return null;
+		}
 	}
 	
 	private void verificarUnidade(Unidade unidade){
@@ -49,5 +103,4 @@ public class UnidadeService{
 			throw new IllegalArgumentException("A unidade nao possui todos os dados.");
 		}
 	}
-	
 }
